+++ elastic.py
@@ -6,10 +6,12 @@
 # Secure Syntax: host = ["https://servername:port", "https://servername2:port"]
 # Unsecure Syntax: host = ["http://servername:port", "http://servername2:port"]
 # Old Syntax: ["servername", "servername2"]
-# NOTE:  If using the old syntax must also set the port and scheme entries.
+# NOTE 1:  If using the old syntax must also set the port and scheme entries.
+# NOTE 2: The old syntax has been deprecated and will be removed in a future version.
 host = ["https://HOST_NAME1:9200", "https://HOST_NAME2:9200"]
 # Elasticsearch port - default is port 9200
-# NOTE: Not required if using new syntax for host entry.
+# NOTE 1: Not required if using new syntax for host entry.
+# NOTE 2: The port entry has been deprecated and will be removed in a future version.
 port = 9200
 # User name
 # If no user name is required then use None
@@ -23,6 +25,7 @@ japd = None
 # Example: ssl_client_ca = "/etc/elasticsearch/certs/ca.pem"
 ssl_client_ca = None
 # Type of connection when using SSL connection.
-# NOTE: Not required if using new syntax for host entry.
+# NOTE 1: Not required if using new syntax for host entry.
+# NOTE 2: The scheme entry has been deprecated and will be removed in a future version.
 scheme = "https"